Revert "Rename to cardListPropTypes"
This reverts commit 1be91e2984e339b4a723e6d0195e0d90500328a6.

# Conflicts:
#	src/pages/settings/Payments/PaymentsPage/paymentsPagePropTypes.js
